# api/nodejs-server/api/swagger.yaml
swagger: "2.0"
info:
  description: "Provides a RESTful API for interacting with Smartsplit. Documentation is written in both French and English."
  version: "v1"
  title: "Smartsplit API"
  x-origin:
  - format: "yaml"
    url: "https://raw.githubusercontent.com/iptoki/smartsplit/develop/api/swagger.yaml"
  x-providerName: "smartsplit.org"
  x-logo:
    url: "http://smartsplit.org/wp-content/uploads/2019/05/Logo-v02-005-noir.png"
#host: api.smartsplit.org
basePath: "/v1"
...
schemes:
- "https"
- "http"
consumes:
- "application/json"
produces:
- "application/json"
security:
- jwt: []
paths:
  /auth:
    post:
      tags:
      - "Authentication Details"
      description: "Get a JWT token for the rest of the requests"
      operationId: "postAuth"
      parameters:
      - in: "body"
        name: "body"
        description: "JSON string containing your password"
        required: true
        schema:
          $ref: "#/definitions/auth"
      responses:
        200:
          description: "OK. JWT returned for use with the rest of the API requests"
          schema:
            $ref: "#/definitions/authResult"
        401:
          description: "Invalid credentials and/or API token"
      x-swagger-router-controller: "AuthenticationDetails"
  /auth/verifyPassword:
    post:
      tags:
      - "Authentication Details"
      description: "Check if the password is valid to access the beta"
      operationId: "postAuthVerifyPassword"
      parameters:
      - in: "body"
        name: "body"
        description: "JSON string containing your authentication details"
        required: true
        schema:
          $ref: "#/definitions/authVerifyPassword"
      responses:
        200:
          description: "Success"
        401:
          description: "Failure. Invalid credentials and/or API token"
      x-swagger-router-controller: "AuthenticationDetails"
  /refreshToken:
    get:
      tags:
      - "Authentication Details"
      description: "Refreshes your current, valid JWT token"
      operationId: "getRefreshToken"
      parameters: []
      responses:
        200:
          description: "OK. JWT returned to use in your subsequent requests"
          schema:
            $ref: "#/definitions/authResult"
        401:
          description: "JWT token is missing or expired"
      security:
      - jwt: []
      x-swagger-router-controller: "AuthenticationDetails"
...
securityDefinitions:
  basicAuth:
    description: "HTTP Basic Authentication"
    type: "basic"
  jwt:
    description: "A JSON Web Token (JWT) used for access control and authorization"
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
  auth:
    properties:
      apiKey:
        type: "string"
      username:
        type: "string"
      password:
        type: "string"
  authResult:
    type: "object"
    properties:
      accessToken:
        type: "string"
        example: "fyJhbGciOiJIUzI1NiIsInR5cCI6IkpXUyJ9.eyJjb21wYW55IjoiRnV0dXJlRWQiLCJzdWIiOjEsImlzcyI6Imh0dHA6XC9cL2Z1dHVyZWVkLmRldlwvYXBpXC92MVwvc3R1ZGVudFwvbG9naW5cL3VzZXJuYW1lIiwiaWF0IjoiMTQyNzQyNjc3MSIsImV4cCI6IjE0Mjc0MzAzNzEiLCJuYmYiOiIxNDI3NDI2NzcxIiwianRpIjoiNmFlZDQ3MGFiOGMxYTk0MmE0MTViYTAwOTBlMTFlZTUifQ.MmM2YTUwMjEzYTE0OGNhNjk5Y2Y2MjEwZDdkN2Y1OTQ2NWVhZTdmYmI4OTA5YmM1Y2QwYTMzZjUwNTgwY2Y0MQ"
        description: "The JWT access token for the current session"
    example:
      accessToken: "fyJhbGciOiJIUzI1NiIsInR5cCI6IkpXUyJ9.eyJjb21wYW55IjoiRnV0dXJlRWQiLCJzdWIiOjEsImlzcyI6Imh0dHA6XC9cL2Z1dHVyZWVkLmRldlwvYXBpXC92MVwvc3R1ZGVudFwvbG9naW5cL3VzZXJuYW1lIiwiaWF0IjoiMTQyNzQyNjc3MSIsImV4cCI6IjE0Mjc0MzAzNzEiLCJuYmYiOiIxNDI3NDI2NzcxIiwianRpIjoiNmFlZDQ3MGFiOGMxYTk0MmE0MTViYTAwOTBlMTFlZTUifQ.MmM2YTUwMjEzYTE0OGNhNjk5Y2Y2MjEwZDdkN2Y1OTQ2NWVhZTdmYmI4OTA5YmM1Y2QwYTMzZjUwNTgwY2Y0MQ"
  authVerifyPassword:
    properties:
      password:
        type: "string"
...